Добавляет компонент Catalog, подключает его к Main. Оптимизирует рендеринг при переключении жанра
